// react-landingpage/src/Services/Interfaces/IGuideCollectionContextData.ts
import { IImageUrlBuilder } from './ImageUrlBuilder';

//todo use component linrary version

//todo move interface
export interface ICollectionObject {
    objectId: string;
    imageId: string;
    title: string;
    imageUrlBuilder: IImageUrlBuilder;
}

export interface ICollectionContextData {
    id: number;
    previewImageSlider: string;
    previewImageCard: string;
    previewImageMediaPlayer: string;
    title: string;
    subtitle: string;
    collectionObjects: ICollectionObject[];
    interval?: number;
    selectedObjectIndex?: number;
    onCollectionObjectSelected?: (index: number) => void;
}